ML_cycle/model_production_deployment/requirements.txt
fastapi
pydantic
prophet
fastapi
uvicorn
pandas
numpy
joblib
wandb
python-dotenv
scikit-learn
streamlit
